+++ src/main/java/com/example/demo/taller1proyecto/repository/MenusRepository.java
@@ -9,20 +9,20 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface MenusRepository extends JpaRepository<Menus, Long> {
+public interface MenusRepository extends JpaRepository<Menus, Integer> {
 
     // @Query("SELECT p.codp FROM Procesos p WHERE p.codm = CAST(:menuId AS
     // Integer)")
-    // List<Procesos> getProcesosByMenuId(@Param("menuId") Long menuId);
+    // List<Procesos> getProcesosByMenuId(@Param("menuId") Integer menuId);
     @Query("SELECT p FROM Procesos p JOIN p.codm m WHERE m.codm = :menuId")
-    List<Procesos> getProcesosByMenuId(@Param("menuId") Long menuId);
+    List<Procesos> getProcesosByMenuId(@Param("menuId") Integer menuId);
 
     // // @Query ("SELECT DISTINCT m FROM Menus m JOIN m.roles r JOIN r.usuarios u"
     // +
     // // "WHERE u.login = ?1 AND u.password = ?2")
     // @Query("SELECT DISTINCT m FROM Menus m JOIN m.roles r JOIN r.usuarios u WHERE
     // u.login = ?1 AND u.password = ?2 AND r.codr = ?3")
-    // List<Menus> findByLoginAndPassword(String login, String password, Long
+    // List<Menus> findByLoginAndPassword(String login, String password, Integer
     // idrol);
 
 }
